refactor: 모듈화, entry 에서 이벤트 등록 함수 호출
