+++ public/javascripts/constants.js
@@ -1,0 +1,4 @@
+export const URL = {
+  questions: "http://localhost:3001/questions",
+  answers: "http://localhost:3001/answers",
+};
